[#2451] Refactor: move filterColumns to column.js utilities file
