+++ client/src/components/visualisation/configMenu/LayerConfigMenu.jsx
@@ -8,7 +8,8 @@ import ToggleInput from '../../common/ToggleInput';
 import Button from '../../common/Button';
 import ColorLabels from './ColorLabels';
 import FilterMenu from './FilterMenu';
-import { filterColumns, checkUndefined, sortAlphabetically } from '../../../utilities/utils';
+import { checkUndefined, sortAlphabetically } from '../../../utilities/utils';
+import { filterColumns } from '../../../utilities/column';
 import * as entity from '../../../domain/entity';
 import { palette } from '../../../utilities/visualisationColors';
 import ConfigMenuSection from '../../common/ConfigMenu/ConfigMenuSection';
