+++ client/src/utilities/utils.js
@@ -38,18 +38,6 @@ export function checkUndefined(object = {}, ...keys) {
   return currentObject[keys[keys.length - 1]];
 }
 
-/*
-# arguments
-columns: an array of objects, each of which has a "type" property
-acceptableTypes: a single string or an array of strings
-
-This could be extended in the future to support filtering on props other than type
-*/
-export function filterColumns(columns = [], acceptableTypes = []) {
-  const types = Array.isArray(acceptableTypes) ? acceptableTypes : [acceptableTypes];
-  return columns.filter(column => types.some(type => type === column.type));
-}
-
 export const sortAlphabetically = (a, b, accessor = val => val) => {
   const nameA = accessor(a).toLowerCase();
   const nameB = accessor(b).toLowerCase();
